+++ sdk/keyvault/azure_security_keyvault_certificates/src/lib.rs
@@ -3,9 +3,11 @@
 
 #![doc = include_str!("../README.md")]
 
+pub mod clients;
+#[allow(unused_imports)]
 mod generated;
 pub mod models;
 mod resource;
 
-pub use generated::{clients, CertificateClient, CertificateClientOptions};
+pub use clients::{CertificateClient, CertificateClientExt, CertificateClientOptions};
 pub use resource::*;